Renaming conin.hmm
