+++ conin/hidden_markov_model/__init__.py
