+++ conin/hidden_markov_model/chmm_algebraic.py
