+++ conin/hidden_markov_model/chmm_oracle.py
