+++ conin/hidden_markov_model/constrained_hmm.py
