+++ conin/hidden_markov_model/hmm.py
