+++ conin/hidden_markov_model/hmm_application.py
