+++ conin/hidden_markov_model/hmm_to_dbn.py
