+++ conin/hidden_markov_model/inference/__init__.py
