+++ conin/hidden_markov_model/learning/__init__.py
