+++ conin/hidden_markov_model/learning/helpers.py
