+++ conin/hidden_markov_model/learning/mcem.py
