+++ conin/hidden_markov_model/tests/__init__.py
