+++ conin/hidden_markov_model/tests/test_application_chmm.py
@@ -6,10 +6,10 @@ import copy
 
 import pyomo.environ as pyo
 
-from conin.hmm import *
+from conin.hidden_markov_model import *
 from conin.oracle_constraints import *
 from conin.constraint import pyomo_constraint_fn
-from conin.hmm.inference import viterbi, a_star, lp_inference, ip_inference
+from conin.hidden_markov_model.inference import viterbi, a_star, lp_inference, ip_inference
 
 
 #
