+++ conin/hidden_markov_model/tests/test_hmm_to_dbn.py
@@ -1,8 +1,8 @@
 import pytest
 
-from conin.hmm import create_dbn_from_hmm
+from conin.hidden_markov_model import create_dbn_from_hmm
 from conin.dynamic_bayesian_network import create_bn_from_dbn
-import conin.hmm.tests.examples as tc
+import conin.hidden_markov_model.tests.examples as tc
 
 
 def test_hmm0_to_dbn():
