+++ conin/hidden_markov_model/tests/test_inference.py
@@ -7,15 +7,15 @@ from conin.util import try_import
 with try_import() as or_topas_available:
     import or_topas.aos as aos
 
-from conin.hmm.inference.recursive_a_star import *
-from conin.hmm import *
-from conin.hmm.inference import a_star
+from conin.hidden_markov_model.inference.recursive_a_star import *
+from conin.hidden_markov_model import *
+from conin.hidden_markov_model.inference import a_star
 from conin.oracle_constraints import *
 
-# from conin.hmm.oracle_chmm import Oracle_CHMM
-from conin.hmm.inference import ip_inference
+# from conin.hidden_markov_model.oracle_chmm import Oracle_CHMM
+from conin.hidden_markov_model.inference import ip_inference
 
-import conin.hmm.tests.examples as tc
+import conin.hidden_markov_model.tests.examples as tc
 
 mip_solver = pyomo.opt.check_available_solvers("gurobi", "highs", "glpk")
 mip_solver = mip_solver[0] if mip_solver else None
