+++ conin/hidden_markov_model/tests/test_pyomo_chmm.py
@@ -1,9 +1,9 @@
 import pytest
 
-import conin.hmm
-import conin.hmm.chmm_algebraic
+import conin.hidden_markov_model
+import conin.hidden_markov_model.chmm_algebraic
 
-import conin.hmm.tests.examples as tc
+import conin.hidden_markov_model.tests.examples as tc
 
 
 class Test_IndexSets1:
@@ -21,7 +21,7 @@ class Test_IndexSets1:
 
     def test_index_sets(self):
         hmm = tc.create_hmm1()
-        data = conin.hmm.chmm_algebraic._create_index_sets(
+        data = conin.hidden_markov_model.chmm_algebraic._create_index_sets(
             hmm=hmm, observed=["o0", "o1", "o0", "o1", "o0"]
         )
         assert data.E == [
@@ -105,7 +105,7 @@ class Test_IndexSets2:
 
     def test_index_sets(self):
         hmm = tc.create_hmm2()
-        data = conin.hmm.chmm_algebraic._create_index_sets(
+        data = conin.hidden_markov_model.chmm_algebraic._create_index_sets(
             hmm=hmm, observed=["o0", "o1", "o0", "o1", "o0", "o2"]
         )
         assert data.E == [
