+++ conin/hidden_markov_model/tests/test_util.py
