+++ conin/inference/tests/test_CFNInference.py
@@ -8,7 +8,7 @@ from conin.inference.CFNInference import (
 )
 import conin.markov_network.examples
 import conin.bayesian_network.examples
-import conin.hmm.tests.examples
+import conin.hidden_markov_model.tests.examples
 import conin.dynamic_bayesian_network.examples
 
 with try_import() as pgmpy_available:
@@ -205,7 +205,7 @@ def Xtest_CFNInference_cancer1_constrained_pgmpy():
 
 @pytest.mark.skipif(not pytoulbar2_available, reason="pytoulbar2 not installed")
 def Xtest_CFNInference_hmm1_test0():
-    pgm = conin.hmm.tests.examples.create_hmm1()
+    pgm = conin.hidden_markov_model.tests.examples.create_hmm1()
     inf = CFNInference(pgm)
     observed = ["o0", "o0", "o1", "o0", "o0"]
     results = inf.map_query(evidence=observed)
@@ -214,7 +214,7 @@ def Xtest_CFNInference_hmm1_test0():
 
 @pytest.mark.skipif(not pytoulbar2_available, reason="pytoulbar2 not installed")
 def Xtest_CFNInference_hmm1_test1():
-    pgm = conin.hmm.tests.examples.create_hmm1()
+    pgm = conin.hidden_markov_model.tests.examples.create_hmm1()
     inf = CFNInference(pgm)
     observed = ["o0", "o1", "o1", "o1", "o1"]
     results = inf.map_query(evidence=observed)
@@ -223,7 +223,7 @@ def Xtest_CFNInference_hmm1_test1():
 
 @pytest.mark.skipif(not pytoulbar2_available, reason="pytoulbar2 not installed")
 def Xtest_CFNInference_hmm1_test2():
-    pgm = conin.hmm.tests.examples.create_hmm1()
+    pgm = conin.hidden_markov_model.tests.examples.create_hmm1()
     inf = CFNInference(pgm)
     observed = {0: "o0", 1: "o0", 2: "o1", 3: "o0", 4: "o0"}
     results = inf.map_query(evidence=observed)
@@ -238,7 +238,7 @@ def Xtest_CFNInference_hmm1_test2():
 
 @pytest.mark.skipif(not pytoulbar2_available, reason="pytoulbar2 not installed")
 def Xtest_CFNInference_hmm1_test3():
-    pgm = conin.hmm.tests.examples.create_hmm1()
+    pgm = conin.hidden_markov_model.tests.examples.create_hmm1()
     inf = CFNInference(pgm)
     observed = {0: "o0", 1: "o1", 2: "o1", 3: "o1", 4: "o1"}
     results = inf.map_query(evidence=observed)
@@ -253,7 +253,7 @@ def Xtest_CFNInference_hmm1_test3():
 
 @pytest.mark.skipif(not pytoulbar2_available, reason="pytoulbar2 not installed")
 def Xtest_CFNInference_chmm1_test0():
-    pgm = conin.hmm.tests.examples.create_chmm1_pyomo()
+    pgm = conin.hidden_markov_model.tests.examples.create_chmm1_pyomo()
     inf = CFNInference(pgm)
     observed = ["o0"] * 15
     results = inf.map_query(evidence=observed)
@@ -278,7 +278,7 @@ def Xtest_CFNInference_chmm1_test0():
 
 @pytest.mark.skipif(not pytoulbar2_available, reason="pytoulbar2 not installed")
 def Xtest_CFNInference_chmm1_test1():
-    pgm = conin.hmm.tests.examples.create_chmm1_pyomo()
+    pgm = conin.hidden_markov_model.tests.examples.create_chmm1_pyomo()
     inf = CFNInference(pgm)
     observed = ["o0"] + ["o1"] * 14
     results = inf.map_query(evidence=observed)
@@ -303,7 +303,7 @@ def Xtest_CFNInference_chmm1_test1():
 
 @pytest.mark.skipif(not pytoulbar2_available, reason="pytoulbar2 not installed")
 def Xtest_CFNInference_chmm1_test2():
-    pgm = conin.hmm.tests.examples.create_chmm1_pyomo()
+    pgm = conin.hidden_markov_model.tests.examples.create_chmm1_pyomo()
     inf = CFNInference(pgm)
     observed = {i: "o0" for i in range(15)}
     results = inf.map_query(evidence=observed)
@@ -328,7 +328,7 @@ def Xtest_CFNInference_chmm1_test2():
 
 @pytest.mark.skipif(not pytoulbar2_available, reason="pytoulbar2 not installed")
 def Xtest_CFNInference_chmm1_test3():
-    pgm = conin.hmm.tests.examples.create_chmm1_pyomo()
+    pgm = conin.hidden_markov_model.tests.examples.create_chmm1_pyomo()
     inf = CFNInference(pgm)
     observed = {0: "o0"}
     for i in range(14):
