+++ conin/inference/tests/test_OptimizationInference.py
@@ -4,7 +4,7 @@ import pyomo.opt
 import conin.markov_network.examples
 import conin.bayesian_network.examples
 import conin.dynamic_bayesian_network.examples
-import conin.hmm.tests.examples
+import conin.hidden_markov_model.tests.examples
 
 from conin.inference.OptimizationInference import (
     IntegerProgrammingInference,
@@ -211,7 +211,7 @@ def test_IntegerProgrammingInference_cancer1_constrained_pgmpy():
 
 @pytest.mark.skipif(not mip_solver, reason="No mip solver installed")
 def test_IntegerProgrammingInference_hmm1_test0():
-    pgm = conin.hmm.tests.examples.create_hmm1()
+    pgm = conin.hidden_markov_model.tests.examples.create_hmm1()
     inf = IntegerProgrammingInference(pgm)
     observed = ["o0", "o0", "o1", "o0", "o0"]
     results = inf.map_query(evidence=observed, solver=mip_solver)
@@ -220,7 +220,7 @@ def test_IntegerProgrammingInference_hmm1_test0():
 
 @pytest.mark.skipif(not mip_solver, reason="No mip solver installed")
 def test_IntegerProgrammingInference_hmm1_test1():
-    pgm = conin.hmm.tests.examples.create_hmm1()
+    pgm = conin.hidden_markov_model.tests.examples.create_hmm1()
     inf = IntegerProgrammingInference(pgm)
     observed = ["o0", "o1", "o1", "o1", "o1"]
     results = inf.map_query(evidence=observed, solver=mip_solver)
@@ -229,7 +229,7 @@ def test_IntegerProgrammingInference_hmm1_test1():
 
 @pytest.mark.skipif(not mip_solver, reason="No mip solver installed")
 def test_IntegerProgrammingInference_hmm1_test2():
-    pgm = conin.hmm.tests.examples.create_hmm1()
+    pgm = conin.hidden_markov_model.tests.examples.create_hmm1()
     inf = IntegerProgrammingInference(pgm)
     observed = {0: "o0", 1: "o0", 2: "o1", 3: "o0", 4: "o0"}
     results = inf.map_query(evidence=observed, solver=mip_solver)
@@ -244,7 +244,7 @@ def test_IntegerProgrammingInference_hmm1_test2():
 
 @pytest.mark.skipif(not mip_solver, reason="No mip solver installed")
 def test_IntegerProgrammingInference_hmm1_test3():
-    pgm = conin.hmm.tests.examples.create_hmm1()
+    pgm = conin.hidden_markov_model.tests.examples.create_hmm1()
     inf = IntegerProgrammingInference(pgm)
     observed = {0: "o0", 1: "o1", 2: "o1", 3: "o1", 4: "o1"}
     results = inf.map_query(evidence=observed, solver=mip_solver)
@@ -259,7 +259,7 @@ def test_IntegerProgrammingInference_hmm1_test3():
 
 @pytest.mark.skipif(not mip_solver, reason="No mip solver installed")
 def test_IntegerProgrammingInference_chmm1_test0():
-    pgm = conin.hmm.tests.examples.create_chmm1_pyomo()
+    pgm = conin.hidden_markov_model.tests.examples.create_chmm1_pyomo()
     inf = IntegerProgrammingInference(pgm)
     observed = ["o0"] * 15
     results = inf.map_query(evidence=observed, solver=mip_solver)
@@ -284,7 +284,7 @@ def test_IntegerProgrammingInference_chmm1_test0():
 
 @pytest.mark.skipif(not mip_solver, reason="No mip solver installed")
 def test_IntegerProgrammingInference_chmm1_test1():
-    pgm = conin.hmm.tests.examples.create_chmm1_pyomo()
+    pgm = conin.hidden_markov_model.tests.examples.create_chmm1_pyomo()
     inf = IntegerProgrammingInference(pgm)
     observed = ["o0"] + ["o1"] * 14
     results = inf.map_query(evidence=observed, solver=mip_solver)
@@ -309,7 +309,7 @@ def test_IntegerProgrammingInference_chmm1_test1():
 
 @pytest.mark.skipif(not mip_solver, reason="No mip solver installed")
 def test_IntegerProgrammingInference_chmm1_test2():
-    pgm = conin.hmm.tests.examples.create_chmm1_pyomo()
+    pgm = conin.hidden_markov_model.tests.examples.create_chmm1_pyomo()
     inf = IntegerProgrammingInference(pgm)
     observed = {i: "o0" for i in range(15)}
     results = inf.map_query(evidence=observed, solver=mip_solver)
@@ -334,7 +334,7 @@ def test_IntegerProgrammingInference_chmm1_test2():
 
 @pytest.mark.skipif(not mip_solver, reason="No mip solver installed")
 def test_IntegerProgrammingInference_chmm1_test3():
-    pgm = conin.hmm.tests.examples.create_chmm1_pyomo()
+    pgm = conin.hidden_markov_model.tests.examples.create_chmm1_pyomo()
     inf = IntegerProgrammingInference(pgm)
     observed = {0: "o0"}
     for i in range(14):
